Move MessageProvider to k9mail module
